Begin moving towards making this package an npm package
Currently, I'm looking into how to resolve the errors with the mixins. it may be a problem with VSCode analyzing the code with the wrong typescript version. Overall, I'd like to figure out what would be the right way to do this with more recent mixin changes to Typescript, but that will take some reorienting to this code that is now a year and a half old